# src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: API Ajmanre
  version: "1.0"
servers:
  - url: http://localhost:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
  - url: https://angular.ajmanre.com:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
...
paths:
...
  /agency/{agencyId}:
    get:
      tags:
        - "agencyId"
      summary: "agency get"
      operationId: "agencyId"
      parameters:
        - in: path
          name : agencyId
          schema:
            type: string
          required: true
          description: "agency details by id"
      responses:
        "200":
          description: "agency pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
...
  /property:
    post:
      tags:
        - "propertyAdd"
      summary: "add property"
      operationId: "propertyPost"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Property"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /property/retrieve:
    post:
      tags:
        - "propertyRetrieve"
      summary: "property retrieve"
      operationId: "propertyRetrieve"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertySpec"
      responses:
        "200":
          description: "agency pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyPage"
  /property/get/{propertyId}:
    get:
      tags:
        - "propertyId"
      summary: "property get"
      operationId: "propertyId"
      parameters:
        - in: path
          name : propertyId
          schema:
            type: string
          required: true
          description: "property details by id"
      responses:
        "200":
          description: "agent pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /property/{id}:
    put:
      tags:
        - "propertyEdit"
      summary: "edit property"
      operationId: "propertyPut"
      parameters:
        - in: path
          name : id
          schema:
            type: string
          required: true
          description: "update property details by id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Property"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /property/my/listing:
    get:
      tags:
        - "myPropListing"
      summary: "my property posted by me list"
      operationId: "myPropListing"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      responses:
        "200":
          description: "my property posted by me list"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
  /property/agency/listing:
    get:
      tags:
        - "agencyPropListing"
      summary: "agency property posted by me list"
      operationId: "agencyPropListing"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      responses:
        "200":
          description: "agency property posted by me list"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
  /property/listing/{agencyId}:
    get:
      tags:
        - "propListingByAgency"
      summary: "property listing by agency"
      operationId: "propListingByAgency"
      parameters:
        - in: path
          name : agencyId
          schema:
            type: string
          required: true
          description: "property listing by agency id"
      responses:
        "200":
          description: "agency property posted by me list"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: "object"
      required:
        - message
      properties:
        message:
          type: "string"
          description: "Message"
        identifier:
          type: "string"
          description: "If identifier has to be in response"
...
    Address:
      type: "object"
      required:
        - adressLine1
        - city
      properties:
        adressLine1:
          type: "string"
        adressLine2:
          type: "string"
        area:
          type: "string"
        city:
          type: "string"
        country:
          type: "string"
        zip:
          type: "string"
    Contact:
      type: "object"
      required:
        - name
      properties:
        name:
          type: "string"
        tel:
          type: "string"
        mobile:
          type: "string"
        email:
          type: "string"
    Agency:
      type: "object"
      required:
        - id
        - name
      properties:
        id:
          type: "string"
        updatedAt:
          type: "string"
          format: "date-time"
        name:
          type: "string"
        tel:
          type: "string"
        facebook:
          type: "string"
          description: "facebook page"
        insta:
          type: "string"
          description: "instagram profile id"
        adresses:
          type: "array"
          description: "addresses of agency"
          items:
            $ref: "#/components/schemas/Address"
        contacts:
          type: "array"
          description: "contacts of agency"
          items:
            $ref: "#/components/schemas/Contact"
...
    Source:
      type: "object"
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: source identifier
        name:
          type: string
          description: key value of the  source
    Property:
      type: "object"
      required:
        - id
        - identifier
        - title
      properties:
        id:
          type: string
          description: source identifier
        updatedAt:
          type: "string"
          format: "date-time"
        identifier:
          type: string
          description: unique identifer
        title:
          type: string
          description: property title
          example: 1BHK 2Baths
        status:
          type: string
          example: Available
        statusId:
          type: string
          example : xxxxxxxxxxxxxxxxxxxxxxxx
        propFor:
          type: string
          example: Rent
        propForId:
          type: string
          example : xxxxxxxxxxxxxxxxxxxxxxxx
        type:
          type: string
          example : Residential
        typeId:
          type: string
          example : xxxxxxxxxxxxxxxxxxxxxxxx
        subType:
          type: string
          example : Apartment
        rentingType:
          type: string
          example: Daily
        rentingTypeId:
          type: string
          example : xxxxxxxxxxxxxxxxxxxxxxxx
        location:
          type: string
          example : Al Jerf Industrial 3
        locationId:
          type: string
          example : xxxxxxxxxxxxxxxxxxxxxxxx
        tags:
          type: array
          items:
            type: string
          example : ['RESALE']
        description:
          type: string
          example : description for detailing the property
        price:
          type: number
          format: double
          description : Sale or Rent Price
          example : 1.12
        sqFt:
          type: number
          format: double
          example : 430
        sqMt:
          type: number
          format: double
          example : 39.95
        buildupArea:
          type: number
          format: double
          example : 39.95
        totalLandArea:
          type: number
          format: double
          example : 39.95
        bathrooms:
          type: integer
          example : 2
        bedrooms:
          type: integer
          example : 1
        yearBuilt:
          type: string
          example: 2016
        rooms:
          type: integer
          example : 1
        parking:
          type: integer
          example : 1
        garage:
          type: integer
          example : 1
        garageSize:
          type: number
          format: double
          example : 100
        availabilityDate:
          type: "string"
          format: "date-time"
          example : 2022-06-09T00:00:00.00Z
        features:
          type: array
          items:
            type: string
          example : ['READY TO MOVE']
        genFeatures:
          type: array
          items:
            type: string
          example : ['Laundry Room']
        nearby:
          type: array
          items:
            type: string
          example : ['Laundry Room']
        tel:
          type: string
          example : 455467898
        phoneNo:
          type: string
          example: 987655578
        address:
          $ref: "#/components/schemas/Address"
        floorPlan:
          $ref: "#/components/schemas/FloorPlan"
        project:
          $ref: "#/components/schemas/Project"
        disclaimer:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/Fileo"
          description: files for property
        assocWith:
          $ref: "#/components/schemas/Source"
        managedBy:
          $ref: "#/components/schemas/Source"
    Project:
      type: "object"
      required:
        - name
      properties:
        name:
          type: string
          description: name of the project
        overview:
          type: string
          description: project short overview
    FloorPlan:
      type: "object"
      required:
        - title
      properties:
        title:
          type: string
          description: title of the floor
        bedroom:
          type: integer
        bathroom:
          type: integer
        price:
          type: number
          format: double
          description : Sale or Rent Price
        postFixPrice:
          type: string
        planSize:
          type: number
          format: double
        planImage:
          $ref: "#/components/schemas/Fileo"
        description:
          type: string
          description: floor short overview
    PropertyPage:
      type: "object"
      required:
        - total
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
          description: data elements list in page
        total:
          type: integer
          format: int64
          description: total elements in the list
    PropertySpec:
      type: "object"
      properties:
        types:
          type: array
          items:
            $ref: "#/components/schemas/TypeRel"
          description: property types
        statusIds:
          type: "array"
          description: "status ids"
          items:
            type: string
        locationIds:
          type: "array"
          description: "location ids"
          items:
            type: string
        bedroom:
          type: integer
          description: no of bedroom
    TypeRel:
      type: "object"
      required:
        - typeId
        - type
        - subType
      properties:
        typeId:
          type: string
        type:
          type: string
        subType:
          type: string
...
    Fileo:
      type: "object"
      required:
        - link
      properties:
        link:
          type: string
          description: downloadable link
        name:
          type: string
          description: file name
...
security:
  - bearerAuth: []  